requirements.txt
keras==2.1.6
opencv-python
pillow
scikit-image
sklearn